refactor(IDisplay) optimize type

+++ src/traffic_light/IDisplay.h
@@ -3,5 +3,5 @@
 class IDisplay {
   public:
     virtual void begin() = 0;
-    virtual void show(int value) = 0;
+    virtual void show(uint8_t value) = 0;
 };
